refactor: update authentication mechanism in test plan script Refactor the test plan generation script to remove reliance on CLERK_SESSION_TOKEN and instead utilize DEV_AUTH_USER_ID or COOKIE_HEADER for local authentication. This change simplifies the setup process for local development and enhances flexibility in testing.
Changes:
- Remove loading of CLERK_SESSION_TOKEN from .env.local
- Introduce read_env_value function to read environment variables
- Validate local auth setup using DEV_AUTH_USER_ID or COOKIE_HEADER
- Update curl commands to use COOKIE_HEADER for authentication
- Modify usage instructions in the script comments
- Remove unused notImplemented and methodNotAllowed functions from response module
- Delete unused API routes and related tests for enhance-content and templates
