--- src/app/api/v1/ai/enhance-content/route.ts
@@ -1,27 +1,0 @@
-import { withAuthAndRateLimit, withErrorBoundary } from '@/lib/api/auth';
-import { notImplemented } from '@/lib/api/response';
-
-/**
- * POST /api/v1/ai/enhance-content (future)
- *  Purpose: Apply targeted AI enhancements to existing plan content WITHOUT full regeneration.
- *  Potential enhancement types:
- *    - 'improve_descriptions' (rewrite module/task descriptions for clarity)
- *    - 'add_practice_tasks' (append tasks focused on hands-on learning)
- *    - 'tighten_scope' (reduce total volume based on updated weekly hours)
- *
- *  Flow (pending design):
- *    1. Validate enhancement_type & optional scope (moduleId / taskId).
- *    2. Insert enhancement job row (plan_enhancements) with status='pending'.
- *    3. Enqueue worker job performing selective modifications (version fields or patch table).
- *    4. Return 202 { enhancementId }.
- *
- *  Open questions:
- *    - Should enhancements be versioned & revertible per task/module?
- *    - Conflict resolution if regeneration occurs mid-enhancement.
- *
- *  MVP decision: Defer implementation until baseline plan lifecycle is stable.
- */
-
-export const POST = withErrorBoundary(
-  withAuthAndRateLimit('aiGeneration', async () => notImplemented())
-);
