+++ src/lib/api/response.ts
@@ -1,7 +1,6 @@
 import type { FailureClassification } from '@/lib/types/client';
 
 import { toApiErrorJsonResponse } from '@/lib/api/error-response';
-import type { AppError } from '@/lib/api/errors';
 
 interface JsonOptions {
   status?: number;
@@ -25,22 +24,4 @@ export function jsonError(
   } = {}
 ) {
   return toApiErrorJsonResponse(message, options);
-}
-
-export function notImplemented() {
-  return jsonError('Not Implemented', {
-    status: 501,
-    code: 'NOT_IMPLEMENTED',
-  });
-}
-
-export function methodNotAllowed() {
-  return jsonError('Method Not Allowed', {
-    status: 405,
-    code: 'METHOD_NOT_ALLOWED',
-  });
-}
-
-export function assert(condition: unknown, err: AppError) {
-  if (!condition) throw err;
 }
