+++ tests/unit/api/response.spec.ts
@@ -1,12 +1,5 @@
+import { json, jsonError } from '@/lib/api/response';
 import { describe, expect, it } from 'vitest';
-import {
-  json,
-  jsonError,
-  notImplemented,
-  methodNotAllowed,
-  assert,
-} from '@/lib/api/response';
-import { AppError } from '@/lib/api/errors';
 
 describe('json', () => {
   it('should create JSON response with default status 200', async () => {
@@ -160,93 +153,5 @@ describe('jsonError', () => {
 
     const body = await response.json();
     expect(body.details).toEqual({ count: 0, enabled: false });
-  });
-});
-
-describe('notImplemented', () => {
-  it('should return 501 status', async () => {
-    const response = notImplemented();
-    expect(response.status).toBe(501);
-  });
-
-  it('should return correct error message and code', async () => {
-    const response = notImplemented();
-    const body = await response.json();
-
-    expect(body).toEqual({
-      error: 'Not Implemented',
-      code: 'NOT_IMPLEMENTED',
-    });
-  });
-});
-
-describe('methodNotAllowed', () => {
-  it('should return 405 status', async () => {
-    const response = methodNotAllowed();
-    expect(response.status).toBe(405);
-  });
-
-  it('should return correct error message and code', async () => {
-    const response = methodNotAllowed();
-    const body = await response.json();
-
-    expect(body).toEqual({
-      error: 'Method Not Allowed',
-      code: 'METHOD_NOT_ALLOWED',
-    });
-  });
-});
-
-describe('assert', () => {
-  it('should not throw when condition is true', () => {
-    const error = new AppError('Test error');
-    expect(() => assert(true, error)).not.toThrow();
-    expect(() => assert(1, error)).not.toThrow();
-    expect(() => assert('value', error)).not.toThrow();
-    expect(() => assert({}, error)).not.toThrow();
-  });
-
-  it('should throw the provided error when condition is false', () => {
-    const error = new AppError('Test error', { status: 400 });
-    expect(() => assert(false, error)).toThrow(error);
-  });
-
-  it('should throw when condition is null', () => {
-    const error = new AppError('Null value');
-    expect(() => assert(null, error)).toThrow(error);
-  });
-
-  it('should throw when condition is undefined', () => {
-    const error = new AppError('Undefined value');
-    expect(() => assert(undefined, error)).toThrow(error);
-  });
-
-  it('should throw when condition is 0', () => {
-    const error = new AppError('Zero value');
-    expect(() => assert(0, error)).toThrow(error);
-  });
-
-  it('should throw when condition is empty string', () => {
-    const error = new AppError('Empty string');
-    expect(() => assert('', error)).toThrow(error);
-  });
-
-  it('should preserve error properties', () => {
-    const error = new AppError('Custom error', {
-      status: 403,
-      code: 'FORBIDDEN',
-      details: { userId: '123' },
-    });
-
-    try {
-      assert(false, error);
-      // Should not reach here
-      expect(true).toBe(false);
-    } catch (err) {
-      expect(err).toBe(error);
-      expect((err as AppError).status()).toBe(403);
-      expect((err as AppError).code()).toBe('FORBIDDEN');
-      expect((err as AppError).details()).toEqual({ userId: '123' });
-    }
   });
 });
